Major refactorization + stability enhancements
In particular, sometimes ARPACK decides that one of the projectors is the
optimum choice for the site vector because the eigenvalue zero is optimum.
Obviously this is *not* what we want, so I monitor for this case and skip the
site if this is what happens.

Also, the scripts are now a little bit smarter:  they repeat sweeps until the
energy converges.