Fixed several bugs.
I was wrong to have the repeated sweeper return the old chain rather than the
new chain, since you only want to do that if the two chains have different
bandwidths.  However, this actually uncovered some other bugs, namely that the
first site tensor of the initial state is unnormalized and thus you get bad
expectations, and also it is not projected so it might not be orthogonal to the
overlap vectors.  Both of these problems had been masked by the fact that they
are both fixed on the very first step of the calculation, but it is still a bad
idea to ever create a chain that is in an inconsistent/incorrect state.