Reversed indices in diagrams and Python tests to match Fortran ordering.
Constantly reversing the indices in my head in order to link them to the indices
in the Fortran code was just becoming too confusing for my poor brain.  :-)

Besides which all of the number crunching happens in Fortran so there isn't
much point in even thinking of them in terms of the C ordering.