Fixed bugs, used "amplification" hack to improve convergence.
The hack replaces the strategy of mapping vectors in the projector space to zero
with a strategy of treating them as eigenvectors with some amplification as the
eigenvalue;  this way the optimizer will not get confused and pick an
eigenvector in the projector space when the lowest eigenvalue in the orthogonal
space is near or just above zero.

It is a "hack" because at the moment this amplification is hard-coded to 1,
since this was optimal for eliminating warnings in my test suite.