Lots of changes.
*) Added more sanity checks, which uncovered a problem in the way that the
projector overlap was computed.

*) Improved the way the projector overlap was computed; it now uses *all* of the
projectors, rather than just the effective projector basis (which might be
smaller).  This allows it to stay constant even as the number of ranks in the
effective projector basis changes.

*) Added a special-case in the main algorithm where it runs strategy 1 if
strategies 2 or 3 fail to converge, since strategy 1 always converges.  This
is currently triggered when the full space dimensions is less than the
number of iterations, but this might be too low of a threshold since strategy 1
could possibly be very expensive.

*) Fixed a bug in the test script.