Migrated the test suite from Illuminate to CTest.
This has the advantage of removing the dependency on Illuminate and allowing me
to leverage the features of the CTest framework in the future.

The downside is that the process of building and running the tests is slower
and the output doesn't look as nice;  hopefully this will be offset in the
future by other features that CTest offers.