Ripped out use of the dimension classes except in constructor arguments.
It turns out that in places other than constructor arguments they were
actually not protecting me from anything since nearly all of the time
I had to extract the integers from them anyway and so they added
inconvenience for nothing in return.

On the other hand, they continue to be useful for constructor arguments
to make sure that passing dimensions in the wrong order is a compile-
time error rather than a runtime error.