Starting rewrite of the C++ layer in Python.  (explained below)
Unfortunately, writing the driver code in C++ has resulted in a system
that is surprisingly brittle.  The code stopped working reliability
for no apparent reason sometime in the last couple of months, and I
have had a hard time getting it to even *compile* in other
environments due to the vagaries of various C++ compilers.  As much
as I hate to rewrite the glue layer *again*, I think that having it be
in Python will result in a system that is much easier to maintain and
deploy to end-users.

The reason why I am choosing Python instead of some other language is
because of the numpy/scipy infrastructure, which comes with
functionality for mixing Python and Fortran code as well as arpack
built-in, which solves a couple of the irritating challenges of
getting this code to build and deploy right off the bat.