Simplified Grids:  now their side boundaries must be simple vectors.
There is an important theoretical reason for this, as well as an
important practical reason.  The theoretical reason is that this
better matches the Buchi automata boundary condition that the
automata must visit a given state or states an infinite number of
times.  The practical reason is that we want to minimize the
dependence of the grid on the boundaries so the loss of information
caused by the compression steps does not completely break the grid.