Migrated project away from Leo, replacing sentinels with explicit fold markers.
The reason for this is that clearly the Leo project is never going to
come up with a distribution that is easy to download and install on
all platforms, so the only one benefiting from the sentinels is me.
Explicit fold markers --- that is, {{{ and }}} --- are by contrast,
fairly standard, and thus benefit people who are using a range of
the major, widely used text editors.