Rebuilding the code from the ground up.
There are two reasons why I am doing this.  First, because my
plans for the 2D environment have changed significantly, and
second, because I want to build in from the beginning an
abstraction over the way that data is stored.  These changes
are far-reaching enough that I would rather start from a
clean slate so that I can pull in old code and get it working
one chunk at a time rather than everywhere simultaneously.